Fix initialized origin not being set as default
When initializing Blizzard with a provided origin it is not correctly
set as the default origin. Calling the API always requires an origin for
each call, otherwise it falls back to "us".

Update endpoints.js to set the valid regional key as origin in the
returned object.

Update blizzard.js to set defaults.origin to the value returned from the
default endpoint during initialization. Update get() to use
defaults.origin if no origin is provided in the function parameters.

Update endpoints.test.js.snap to include the origin being returned in
the endpoint object.

Update blizzard.test.js to test several scenarios of origin being
provided and not provided when calling the get() function.